added proper database migration for ver2 of database

+++ lib/ui/CalculationPages/bmrContainer.dart
@@ -22,7 +22,7 @@ class BmrContainer extends StatelessWidget {
 
   const BmrContainer(this.context,this.profileModel);
 
-  double calculateBMI(int weight, int height) {
+  double calculateBMI(double weight, double height) {
     return (703 * (weight / (height * height))).toDouble();
   }
